fix: move handler functions to hander package
